devide each schemas to file
