+++ app/client/components/story/StoryList.tsx
@@ -3,16 +3,6 @@ import StoryItem from './StoryItem';
 import { useQuery } from '@apollo/client';
 import { GET_STORIES } from '../../query';
 
-// Todo: Move to graphql folder later
-// const GET_STORIES = gql`
-//   query Query {
-//     getStories {
-//       id
-//       name
-//     }
-//   }
-// `;
-
 const StoryList = () => {
   const { loading, error, data } = useQuery(GET_STORIES);
   if (loading) {
