+++ app/client/query/editStory.ts
@@ -1,0 +1,15 @@
+import { gql } from '@apollo/client';
+
+export const EDIT_STORY = gql`
+  mutation Mutation($targetId: ID!, $newName: String!) {
+    renameStory(targetId: $targetId, newName: $newName) {
+      code
+      success
+      message
+      story {
+        id
+        name
+      }
+    }
+  }
+`;
