+++ app/client/query/index.ts
@@ -1,1 +1,4 @@
 export * from './getStories';
+export * from './makeStory';
+export * from './deleteStory';
+export * from './editStory';
