+++ app/server/src/index.ts
@@ -1,4 +1,4 @@
-const { ApolloServer, gql } = require('apollo-server');
+const { ApolloServer } = require('apollo-server');
 import { typeDefs } from './schema';
 import { resolvers } from './resolver';
